move Colls.scala to collection package;
